requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
Django==3.2.6
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
multidict==5.1.0
pycodestyle==2.7.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
slackclient==2.9.3
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
